refactor(team): extract task lifecycle into broker_tasks.go
Slice 10. Largest slice in the series — moves the task lifecycle code
(~1,746 lines) into a new broker_tasks.go:
- Task helper funcs: taskNeedsLocalWorktree, taskBlockReasonLooksLikeWorkspaceWriteIssue,
  rejectFalseLocalWorktreeBlock, taskRequiresExclusiveOwnerTurn,
  taskStatusConsumesExclusiveOwnerTurn, taskChannelCandidateOwnerAllowed
- Worktree/lane helpers: syncTaskWorktreeLocked, reusableDependencyWorktreeLocked,
  activeExclusiveOwnerTaskLocked, queueTaskBehindActiveOwnerLaneLocked,
  preferredTaskChannelLocked
- HTTP handlers: handleTasks, handleAgentLogs, handleGetTasks, handlePostTask,
  handleTaskPlan, handleTaskAck
- Notifications: postTaskReassignNotificationsLocked,
  postTaskCancelNotificationsLocked, postTaskDMLocked, isDMTargetSlug,
  dedupeReassignTags
- Lifecycle: BlockTask, ResumeTask, EnsureTask, EnsurePlannedTask,
  AppendTaskDetail, appendTaskDetailLocked, hasUnresolvedDepsLocked,
  unblockDependentsLocked
- Reuse: taskReuseMatch + helpers, hasScopedTaskIdentity,
  taskOwnerMatches, scopedTaskIdentityMatches, findReusableTaskLocked,
  isTerminalTeamTaskStatus
- plannedTaskInput type

Relocates 36 task-themed tests from broker_test.go into a new
broker_tasks_test.go (lifecycle, reassign, cancel, plan, post, block,
resume, unblock, reuse, worktree assignment, ledger/review,
in-flight, intervals).

Adds two characterization tests:
- TestHasUnresolvedDepsLocked_OnlyDoneCounts (cancelled/missing deps
  count as unresolved; only Status=done resolves)
- TestTaskReuseMatch_HasScopedIdentity (whitespace-only identity
  fields don't register as scoped)

Total package coverage: 63.2% (vs 62.9% baseline floor; -0.1pp from
prior slice's 63.3%, still well above floor).
